pyproject.toml
[project]
name = "currency-exchange"
version = "0.1.0"
description = ""
authors = [
    {name = "Woolf",email = "woolf.no1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy[asyncio] (>=2.0.39,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "aiohttp (>=3.11.14,<4.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
asynctest = "^0.13.0"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"